doc_generator/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
google-generativeai==0.3.0
python-multipart==0.0.6
pydantic==2.1.0
jinja2==3.1.2